move `proxy` module to root in `mcp-guardian-core`
