+++ mcp-guardian-core/src/lib.rs
@@ -1,6 +1,10 @@
 pub mod config;
 pub mod dirs;
 pub mod guard_profile;
+pub mod message;
+pub mod message_approval;
+pub mod message_interceptor;
 pub mod proxy;
+pub mod request_cache;
 
 static APP_NAME: &str = "mcp-guardian";
