+++ mcp-guardian-core/src/message.rs
